requirements.txt
pytest==8.2.1
selene>=2.0.0rc6
allure-pytest==2.13.5
python-dotenv==1.0.1
requests==2.32.3
pytest-dotenv==0.5.2
jsonschema==4.22.0
Appium-Python-Client==4.0.0
pydantic==2.7.2